Support TLS with termination past the proxy
This commit makes the proxy addressable by the other services,
so the individual services can communicate with one another via
TLS by addressing each other through the proxy.

There are some services (kratos Admin API, hydra Admin API) which
are not exposed via the proxy. Communication with these is still
direct and via HTTP. If needed, we could allow routing to these
via the proxy too, and apply some CIDR-based filtering to prevent
exposing them outside the docker network. Or introduce an
intermediate proxy.

But for now, the intention is that this communication with the
Admin APIs will always be on the same VM, so TLS is not vital.

Note: Services expect a verifiable certificate. Self-signed certs
do not work yet.