Add mapping of roles('profiles')->permissions('privileges')
Introduces a config file that can be validated, which lets you
map a named role to an array of permissions.

The named roles should correspond with a value available in the
enum for "security-groups" claim in the new kratos identity json.

Also adds a claim for "sofi-data-clearance" which corresponds to
the DataEjerrettigheder FVST/SSI have described in
Brugerprofiler.xlsx