feat: Convert date object to specified string when downloading
Data is aggregated in the view, and it is not straightforward to figure out where which data is coming from, some date is aggregated and other date is plain and simple inserted into a given collection.
Ideally, we would like to have the same data in the view, in the download and in the database, but this is not achievable. Date should be fully qualified dates in the DB, to ensure consistency with timezone. However, it would be nice to have the same data in the view and download file, but the view converts data in different components and the download downloads from the source...

#104224